pyproject.toml
[project]
name = "miruku"
version = "0.1.0"
description = ""
authors = [
    {name = "EvseyKI"}
]
requires-python = ">=3.12,<3.15"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "gql[aiohttp] (>=4.0.0,<5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "qdrant-client (>=1.17.0,<2.0.0)",
    "langchain-core (>=1.2.17,<2.0.0)",
    "python-telegram-bot (>=22.6,<23.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "streamlit (>=1.55.0,<2.0.0)",
    "ddgs (>=9.11.2,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipywidgets (>=8.1.8,<9.0.0)",
    "black (>=26.3.0,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "trackio (>=0.18.0,<0.19.0)",
]
